+++ RELEASE_NOTES.md
@@ -2,7 +2,7 @@
 
 ## Summary
 
-<!-- Here goes a general summary of what this release is about -->
+This release adds a new Assets API based configuration loader, introduces helpers to merge microgrid configs, and updates PV curtailability behavior to support unspecified values.
 
 ## Upgrading
 
@@ -10,7 +10,9 @@
 
 ## New Features
 
-* Add start and end time to microgrid configs.
+* Added `MicrogridConfig.load_configs_from_assets_api(...)` to load microgrid metadata (latitude/longitude) from the Assets API and optionally populate formulas from the component graph.
+* Added `merge_microgrid_configs(...)` for deep-merging two `MicrogridConfig` objects where override values take precedence and `None` does not overwrite base values.
+* Added `merge_config_maps(...)` for merging two dictionaries of microgrid configs by microgrid ID.
 
 ## Bug Fixes